resume-ai/resume-ai/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pinecone-client
pdfplumber
reportlab
pydantic
python-multipart
